+++ src/couchdb/replicator-setup-document.ts
@@ -11,25 +11,49 @@ import { asHealingObject } from '../cleaners/as-healing-object'
 import { ReplicatorDocument, ReplicatorEndpoint } from './replicator-document'
 import { DatabaseSetup } from './setup-database'
 
+/**
+ * Configures which databases a CouchDB cluster contains,
+ * and who it replicates with.
+ *
+ * The various filters wildcard matching with a trailing `*`,
+ * For example, 'logs-*' matches 'logs-2019' and 'logs-2020'.
+ */
 interface ReplicatorCluster {
-  // The base URI to connect to:
-  url: string
+  /**
+   * The base URI to connect to.
+   * @deprecated Use `CouchPool` to hold credentials
+   */
+  url?: string
 
-  // The base64 "username:password" pair used to authenticate.
-  // To get this, run `btoa('username:password')` in a JS console:
+  /**
+   * The base64-encoded "username:password" used for authentication.
+   * To generate this, run `btoa('username:password')` in a browser console.
+   * @deprecated Use `CouchPool` to hold credentials
+   */
   basicAuth?: string
 
-  // Database names that exist on this cluster.
-  // Use an ending '*' for wildcard matching.
-  // For example, 'logs-*' matches things like 'logs-2019' or 'logs-2020'.
-  exclude?: string[] // Do not create or replicate these
-  include?: string[] // Create and replicate these
-  localOnly?: string[] // Create these but don't replicate them
+  /**
+   * List of databases to exclude from creation or replication.
+   * Supports wildcards and tags. Defaults to `['#exclude']`.
+   */
+  exclude?: string[]
 
-  // Cluster names to replicate with.
-  // Use an ending '*' for wildcard matching.
-  // For example, 'logs-*' matches things like 'logs-eu' or 'logs-us'.
+  /**
+   * List of databases to create and replicate.
+   * Supports wildcards and tags. Defaults to `[*]`
+   */
+  include?: string[]
+
+  /**
+   * List of databases to create but not replicate.
+   * Supports wildcards and tags.
+   */
+  localOnly?: string[]
+
+  /** Cluster names to replicate with. */
   pullFrom: string[]
+
+  /** Cluster names to replicate with. */
   pushTo: string[]
 }
 
@@ -128,23 +152,31 @@ export function makeReplicatorDocuments(
     if (!replicated) continue
 
     if (includesName(cluster.pullFrom, remoteName)) {
-      documents[`${name}.from.${remoteName}`] = {
-        continuous: true,
-        create_target: false,
-        owner: currentUsername,
-        source: makeEndpoint(remoteCluster, name),
-        target: makeEndpoint(cluster, name)
+      const source = makeEndpoint(remoteCluster, name)
+      const target = makeEndpoint(cluster, name)
+      if (source != null && target != null) {
+        documents[`${name}.from.${remoteName}`] = {
+          continuous: true,
+          create_target: false,
+          owner: currentUsername,
+          source,
+          target
+        }
       }
     }
 
     if (includesName(cluster.pushTo, remoteName)) {
-      documents[`${name}.to.${remoteName}`] = {
-        continuous: true,
-        create_target: true,
-        create_target_params: options,
-        owner: currentUsername,
-        source: makeEndpoint(cluster, name),
-        target: makeEndpoint(remoteCluster, name)
+      const source = makeEndpoint(cluster, name)
+      const target = makeEndpoint(remoteCluster, name)
+      if (source != null && target != null) {
+        documents[`${name}.to.${remoteName}`] = {
+          continuous: true,
+          create_target: true,
+          create_target_params: options,
+          owner: currentUsername,
+          source,
+          target
+        }
       }
     }
   }
@@ -212,11 +244,14 @@ function includesName(list: string[], name: string): boolean {
 function makeEndpoint(
   cluster: ReplicatorCluster,
   db: string
-): ReplicatorEndpoint {
+): ReplicatorEndpoint | undefined {
+  if (cluster.url == null) return
   const url = `${cluster.url.replace(/[/]$/, '')}/${db}`
-  return cluster.basicAuth == null
-    ? url
-    : { url, headers: { Authorization: `Basic ${cluster.basicAuth}` } }
+  if (cluster.basicAuth == null) return url
+  return {
+    url,
+    headers: { Authorization: `Basic ${cluster.basicAuth}` }
+  }
 }
 
 const asClusterRowRaw = asObject({