+++ src/couchdb/setup-database.ts
@@ -5,6 +5,7 @@ import {
   asMaybeExistsError,
   asMaybeNotFoundError
 } from './couch-error-cleaners'
+import { connectCouch, CouchPool } from './couch-pool'
 import {
   clusterHasDatabase,
   makeReplicatorDocuments,
@@ -18,76 +19,116 @@ import { watchDatabase, WatchDatabaseOptions } from './watch-database'
  */
 export interface DatabaseSetup
   extends Pick<WatchDatabaseOptions, 'onChange' | 'syncedDocuments'> {
-  // The database name:
+  /** The database name. */
   name: string
 
-  // Options to pass to CouchDB when creating this database:
+  /** Options to pass to CouchDB when creating this database. */
   options?: DatabaseCreateParams
 
-  // Documents that should exactly match:
+  /** Documents that should exactly match. */
   documents?: { [id: string]: object }
 
-  // Documents that we should create, unless they already exist:
+  /** Documents that we should create, unless they already exist. */
   templates?: { [id: string]: object }
 
-  // Used for filtering, in addition to the wallet name:
+  /** Used for replicator filtering, in addition to the wallet name. */
   tags?: Array<`#${string}`>
 
-  // Deprecated. Adds '#archived' to default tag list:
+  /** @deprecated Adds '#archived' to default tag list. */
   ignoreMissing?: boolean
 
-  // Deprecated. Put this in the options instead:
+  /** @deprecated Put this in the options instead. */
   replicatorSetup?: SyncedDocument<ReplicatorSetupDocument>
 }
 
 export interface SetupDatabaseOptions {
-  // The couch cluster name the current client is connected to,
-  // as described in the replicator setup document.
-  // This controls which databases and replications we create.
-  // Falls back to "default" if missing:
-  currentCluster?: string
-
-  // Describes which database and replications should exist on each cluster:
+  /**
+   * Describes which database and replications should exist
+   * on each cluster
+   */
   replicatorSetup?: SyncedDocument<ReplicatorSetupDocument>
 
-  // The setup routine will subscribe to the changes feed if
-  // the setup includes an `onChange` callback or synced documents.
-  // This option disables watching, performing a one-time sync instead.
+  /**
+   * The setup routine will subscribe to the changes feed if
+   * the setup includes an `onChange` callback or synced documents.
+   * This option disables watching, performing a one-time sync instead.
+   */
   disableWatching?: boolean
 
-  // Logs status messages whenever we write things to Couch:
+  /**
+   * Don't create databases, design documents, or replications,
+   * but do print messages.
+   */
+  dryRun?: boolean
+
+  /**
+   * Do not do any setup or replication work.
+   */
+  syncOnly?: boolean
+
+  /** Logs status messages whenever we write things to Couch. */
   log?: (message: string) => void
 
-  // Logs error messages whenever something goes wrong:
+  /** Use this cluster for watching for changes. */
+  watchCluster?: string
+
+  /** Logs error messages whenever something goes wrong. */
   onError?: (error: unknown) => void
+
+  /**
+   * @deprecated Pass a CouchPool to `setupDatabase` instead.
+   * The couch cluster name the current client is connected to,
+   * as described in the replicator setup document.
+   * This controls which databases and replications we create.
+   * Falls back to "default" if missing:
+   */
+  currentCluster?: string
 }
 
 /**
  * Ensures that the requested database exists in CouchDB.
+ *
+ * The first parameter should be a `CouchPool` object.
+ * Passing anything else is deprecated.
+ *
  * Returns a cleanup function, which removes any background tasks.
  */
 export async function setupDatabase(
-  connectionOrUri: nano.ServerScope | string,
+  poolOrConnection: CouchPool | nano.ServerScope | string,
   setupInfo: DatabaseSetup,
   opts: SetupDatabaseOptions = {}
 ): Promise<() => void> {
   const { name, onChange, syncedDocuments = [] } = setupInfo
   const {
-    replicatorSetup = setupInfo.replicatorSetup,
+    currentCluster = 'default',
     disableWatching = false,
     log = console.log,
+    replicatorSetup = setupInfo.replicatorSetup,
+    syncOnly = false,
+    watchCluster,
     onError = error => {
       log(`Error while maintaining database "${name}": ${String(error)})`)
     }
   } = opts
-  const connection =
-    typeof connectionOrUri === 'string'
-      ? nano(connectionOrUri)
-      : connectionOrUri
+
+  const pool =
+    typeof poolOrConnection === 'string' || 'relax' in poolOrConnection
+      ? connectCouch(currentCluster, { [currentCluster]: poolOrConnection })
+      : poolOrConnection
 
   // Run the setup once to ensure the database exists:
-  const db = await doSetup(connection, setupInfo, opts)
-  if (db == null) return () => {}
+  if (!syncOnly) await doSetups(pool, setupInfo, opts)
+  const connection =
+    watchCluster == null ? pool.default : pool.connect(watchCluster)
+  const db = connection.use(name)
+
+  // Should we sync documents?
+  const { exists } = clusterHasDatabase(
+    replicatorSetup?.doc,
+    watchCluster ?? pool.defaultName,
+    setupInfo
+  )
+  if (!exists) return () => {}
 
   // Update or watch synced documents:
   const cleanups: Array<() => void> = []
@@ -104,7 +145,7 @@ export async function setupDatabase(
   if (replicatorSetup != null && !disableWatching) {
     cleanups.push(
       replicatorSetup.onChange(() => {
-        doSetup(connection, setupInfo, opts).catch(onError)
+        doSetups(pool, setupInfo, opts).catch(onError)
       })
     )
   }
@@ -112,39 +153,68 @@ export async function setupDatabase(
   return () => cleanups.forEach(cleanup => cleanup())
 }
 
+async function doSetups(
+  pool: CouchPool,
+  setupInfo: DatabaseSetup,
+  opts: SetupDatabaseOptions
+): Promise<void> {
+  const { replicatorSetup = setupInfo.replicatorSetup } = opts
+
+  for (const clusterName of pool.clusterNames) {
+    // Bail out if the current cluster doesn't have this database:
+    const { exists, replicated } = clusterHasDatabase(
+      replicatorSetup?.doc,
+      clusterName,
+      setupInfo
+    )
+
+    if (exists) {
+      await doSetup(pool, clusterName, setupInfo, opts)
+      if (replicated) {
+        await doReplication(pool, clusterName, setupInfo, opts)
+      }
+    }
+  }
+}
+
 /**
  * Performs the actual work of database setup.
  * This is a one-shot process, and doesn't subscribe to anything.
  */
 async function doSetup(
-  connection: nano.ServerScope,
+  pool: CouchPool,
+  clusterName: string,
   setupInfo: DatabaseSetup,
   opts: SetupDatabaseOptions
-): Promise<DocumentScope<unknown> | undefined> {
+): Promise<void> {
   const { documents = {}, name, options, templates = {} } = setupInfo
-  const {
-    currentCluster,
-    log = console.log,
-    replicatorSetup = setupInfo.replicatorSetup
-  } = opts
+  const { dryRun = false, log = console.log } = opts
+  const prefix = dryRun ? 'dry-run: ' : ''
 
-  // Bail out if the current cluster doesn't have this database:
-  const { exists, replicated } = clusterHasDatabase(
-    replicatorSetup?.doc,
-    currentCluster,
-    setupInfo
-  )
-  if (!exists) return
+  const connection = pool.maybeConnect(clusterName)
+  if (connection == null) return
 
   // Create the database if needed:
   const existingInfo = await connection.db.get(name).catch(error => {
     if (asMaybeNotFoundError(error) == null) throw error
   })
   if (existingInfo == null) {
+    log(`${prefix}Creating database "${name}" on cluster "${clusterName}".`)
+
+    // Bail out (with logs) if this is a dry-run:
+    if (dryRun) {
+      const docs = [...Object.keys(documents), ...Object.keys(templates)]
+      for (const id of docs) {
+        log(
+          `${prefix}Writing document "${id}" in database "${name}" on cluster "${clusterName}".`
+        )
+      }
+      return
+    }
+
     await connection.db.create(name, options).catch(error => {
       if (asMaybeExistsError(error) == null) throw error
     })
-    log(`Created database "${name}"`)
   }
   const db: DocumentScope<unknown> = connection.db.use(name)
 
@@ -156,8 +226,11 @@ async function doSetup(
     })
 
     if (!matchJson(documents[id], rest)) {
+      log(
+        `${prefix}Writing document "${id}" in database "${name}" on cluster "${clusterName}".`
+      )
+      if (dryRun) continue
       await db.insert({ _id, _rev, ...documents[id] })
-      log(`Wrote document "${id}" in database "${name}".`)
     }
   }
 
@@ -169,31 +242,77 @@ async function doSetup(
     })
 
     if (_rev == null) {
+      log(
+        `${prefix}Writing document "${id}" in database "${name}" on cluster "${clusterName}".`
+      )
+      if (dryRun) continue
       await db.insert({ _id, ...templates[id] })
-      log(`Wrote document "${id}" in database "${name}".`)
+    }
+  }
+}
+
+async function doReplication(
+  pool: CouchPool,
+  clusterName: string,
+  setupInfo: DatabaseSetup,
+  opts: SetupDatabaseOptions
+): Promise<void> {
+  const replicatorSetup = mergeReplicatorCredentials(
+    pool,
+    opts.replicatorSetup?.doc ?? setupInfo.replicatorSetup?.doc
+  )
+
+  const connection = pool.maybeConnect(clusterName)
+  if (connection == null) return
+
+  // Figure out the current username:
+  const credential = pool.getCredential(clusterName)
+  const currentUsername =
+    credential?.username ??
+    // We can remove this slow fallback in the next breaking release,
+    // since this will only happen when putting a `nano.ServerScope` into
+    // CouchPool, which is deprecated:
+    (await connection.session()).userCtx.name
+
+  // Set up replication:
+  const documents = makeReplicatorDocuments(
+    replicatorSetup,
+    clusterName,
+    currentUsername,
+    setupInfo
+  )
+  await doSetup(pool, clusterName, { name: '_replicator', documents }, opts)
+}
+
+/**
+ * Adds credentials to a replicator setup document.
+ *
+ * The replicator document can contain credentials, but that's deprecated.
+ * Copy credentials from the CouchPool to the replicator document,
+ * so we remain backwards-compatible.
+ */
+function mergeReplicatorCredentials(
+  pool: CouchPool,
+  doc: ReplicatorSetupDocument | undefined
+): ReplicatorSetupDocument {
+  const merged: ReplicatorSetupDocument = { clusters: {} }
+  if (doc == null) return merged
+
+  for (const name of Object.keys(doc.clusters)) {
+    const row = doc.clusters[name]
+    merged.clusters[name] = row
+
+    // If we have credentials in the pool, use those:
+    const credential = pool.getCredential(name)
+    if (credential != null) {
+      const { url, username, password } = credential
+      const basicAuth =
+        username != null && password != null
+          ? btoa(`${username}:${password}`)
+          : undefined
+      merged.clusters[name] = { ...row, url, basicAuth }
     }
   }
 
-  if (replicated) {
-    // Figure out the current username:
-    const sessionInfo = await connection.session()
-    const currentUsername: string = sessionInfo.userCtx.name
-
-    // Set up replication:
-    await doSetup(
-      connection,
-      {
-        name: '_replicator',
-        documents: makeReplicatorDocuments(
-          replicatorSetup?.doc,
-          currentCluster,
-          currentUsername,
-          setupInfo
-        )
-      },
-      opts
-    )
-  }
-
-  return db
+  return merged
 }